+++ src/main/java/org/stellar/sdk/xdr/CryptoKeyType.java
@@ -14,6 +14,7 @@ import java.io.IOException;
 //      KEY_TYPE_ED25519 = 0,
 //      KEY_TYPE_PRE_AUTH_TX = 1,
 //      KEY_TYPE_HASH_X = 2,
+//      KEY_TYPE_ED25519_SIGNED_PAYLOAD = 3,
 //      // MUXED enum values for supported type are derived from the enum values
 //      // above by ORing them with 0x100
 //      KEY_TYPE_MUXED_ED25519 = 0x100
@@ -24,6 +25,7 @@ public enum CryptoKeyType implements XdrElement {
   KEY_TYPE_ED25519(0),
   KEY_TYPE_PRE_AUTH_TX(1),
   KEY_TYPE_HASH_X(2),
+  KEY_TYPE_ED25519_SIGNED_PAYLOAD(3),
   KEY_TYPE_MUXED_ED25519(256),
   ;
   private int mValue;
@@ -42,6 +44,7 @@ public enum CryptoKeyType implements XdrElement {
       case 0: return KEY_TYPE_ED25519;
       case 1: return KEY_TYPE_PRE_AUTH_TX;
       case 2: return KEY_TYPE_HASH_X;
+      case 3: return KEY_TYPE_ED25519_SIGNED_PAYLOAD;
       case 256: return KEY_TYPE_MUXED_ED25519;
       default:
         throw new RuntimeException("Unknown enum value: " + value);